pyproject.toml
[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project]
name = "mcp-server-snowflake-minimal"
version = "1.0.0"
description = "Minimal Snowflake MCP Server with Object Management, SQL Execution, and Semantic View Querying"
authors = [
    {name = "Snowflake Inc.", email = "noreply@snowflake.com"},
]
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.8.1",
    "mcp[cli]>=1.9.4", 
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "snowflake-connector-python>=3.15.0",
    "snowflake-core>=1.0.0",
    "sqlglot>=27.8.0"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pyright", 
    "pytest",
    "python-dotenv",
    "ruff"
]

[project.scripts]
snowflake-mcp-minimal = "snowflake_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/snowflake_mcp"]

[tool.hatch.metadata]
allow-direct-references = true
